revert parser function signature
switched back to using open file objects as parser function inputs.
using `pathlib.Path` objects was nice, it gave each parser function
control over how to open the file and made working with file names and
extensions easier, but at the cost of added complexity to support
backwards compatibility with old parsers.
The biggest problem with the new implementation was that it couldn't
support parsing data from stdin, which is one of the goals of yasha 5.0