Remove unnecessary environment isolation logic
While writing the Yasha class, i kept constantly in the back of my mind
the need to eventually support rendering multiple templates or a
directory of templates with a single class instance. In order to support
this functionality, I embedded logic in most of the Yasha class methods
to work with isolated environments.

Turns out that was kind of dumb. Forcing every method of a class to
adopt extraneous logic just to support a single uncommon use case is
ineffective, and makes the code harder to maintain.

All the code necessary for template environment isolation should stay
inside the method (not yet implemented) which handles multi-template
rendering